third_party/blink/web_tests/css-parser/image-orientation.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>
assert_valid_value("image-orientation", "none");
assert_valid_value("image-orientation", "from-image");

assert_invalid_value("image-orientation", "0deg");
</script>
